requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.39.3
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.2
idna==3.4
ipykernel==6.22.0
ipython==8.13.1
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
langchain==0.0.153
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.5
openapi-schema-pydantic==1.2.4
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
py-cord==2.4.1
pydantic==1.10.7
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.29.0
six==1.16.0
SQLAlchemy==2.0.12
stack-data==0.6.2
tenacity==8.2.2
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
